Add basic support for embedded types
For example, given:
```javascript
{
  {
    id: 'container',
    type: 'schema',
    resourceFields: {
      restart: {
        type: 'restartPolicy'
      }
    }
  },

  {
    id: 'restartPolicy',
    type: 'schema',
    resourceFields: {
      name: {
        type: 'string'
      },

      maximumRetryCount: {
        type: 'int',
      }
    }
  }
}
```

the `restart` field will now show up as a textarea that you can put JSON in.
It would be nicer if it looked up the properties of a restartPolicy and allowed
you to enter them directly, but nesting complex types is not a terribly simple
change so this will provide a way to set embedded types in the meantime.